fix: change Execution.Price to decimal for financial precision + document orderId mapping
Code quality review fixes:

Critical #1: Type Mismatch - Execution.Price (double) vs decimal
- Changed: src/IBApi.Stub/EWrapper.cs:200
  public double Price → public decimal Price
- Rationale: Avoid binary floating-point precision loss in financial calculations
- Removed unnecessary cast in test: (decimal)execution.Price → execution.Price

Important #2: Document orderId Mapping Simplification
- Added comment in TwsCallbackHandlerTests.cs:440-443
- Documents that test uses simplified mapping: orderId == execution.OrderId.ToString()
- Real implementation (Task #8) must do order_tracking lookup:
  SELECT order_id FROM order_tracking WHERE ibkr_order_id = execution.OrderId
- Handles orphaned executions (orders placed before crash with no tracking record)

Build verification: Still fails as expected (RED phase)
CS0535: OrderEventsRepository does not implement InsertExecutionAsync()

+++ src/IBApi.Stub/EWrapper.cs
@@ -197,7 +197,7 @@ public class Execution
     public string Time { get; set; } = "";
     public string Side { get; set; } = "";
     public decimal Shares { get; set; }
-    public double Price { get; set; }
+    public decimal Price { get; set; }  // decimal for financial precision (not double)
     public string Exchange { get; set; } = "";
     public int PermId { get; set; }
     public int ClientId { get; set; }
